requirements.txt
Django==3.1.1
django-cors-headers==3.8.0
django-utils-six==2.0
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.7.3.2
pytz==2017.3
requests==2.20.1
six==1.11.0
whitenoise==5.2.0
django-allauth==0.45.0
django-heroku==0.3.1
djangorestframework==3.12.4

#djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
